Separate Prometheus Page from Prometheus Metrics
